Refactor keyboardBehaviour for readability
